% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2tag.R
\name{array2tag}
\alias{array2tag}
\title{array2tag}
\usage{
array2tag(dir = NULL, ID = NULL, string_del = NULL, propagated = FALSE, save.txt = FALSE)
}
\arguments{
\item{ID}{Optional argument for specimens IDs. The default (NULL) is the file names.}

\item{dir}{Optional argument. Directory where the *.ascii files are. Default (NULL) is the current working directory.}

\item{string_del}{Optional argument. Indicates pattern in the specimens names to be deleted. Use only if ID = NULL.}

\item{propagated}{Optional argument indicating if the landmarks have been propagated (extra line). The default is FALSE.}

\item{save.txt}{Optional argument. Indicates whether to also save *.txt files with the landmark coordinates.}
}
\value{
This function returns an array of dimensions (p, k, n), in which p is the number of landmarks, k = 3 (number of dimensions), and n is the number of specimens.
}
\description{
This function inputs an array of dimensions p, k, n and saves *tag files
}
\examples{
# my_array <- array(1:10, c(5,3,4))
# dimnames(my_array)[[3]] <- c("spec1", "spec2", "spec3", "spec4")
# array2tag(my_array)
  
}
\author{
Marta Vidal-Garcia
}
