% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsv2array.R
\name{fcsv2array}
\alias{fcsv2array}
\title{fcsv2array}
\usage{
fcsv2array(dir = NULL, ID = NULL, string_del = NULL, propagated = FALSE, save.txt = FALSE)
}
\arguments{
\item{dir}{Optional argument. Directory where the *.ascii files are. Default (NULL) is the current working directory.}

\item{pattern}{Optional argument for specific files with a pattern. E.g. "_semilandmarks.fcsv". The default (NULL) is all the ".fcsv" files in a directory.}

\item{ID}{Optional argument for specimens IDs. The default (NULL) is the file names.}

\item{string_del}{Optional argument. Indicates pattern in the specimens names to be deleted. Use only if ID = NULL.}

\item{save.txt}{Optional argument. Indicates whether to also save *.txt files with the landmark coordinates.}
}
\value{
This function returns an array of dimensions (p, k, n), in which p is the number of landmarks, k = 3 (number of dimensions), and n is the number of specimens.
}
\description{
This function inputs *fcsv files (3DSlicer) into an array of dimensions p, k, n
}
\examples{
# array <- fcsv2array()
# If we have other *.fcsv files in another directory and with the suffix "skull"
# For example: "spec1_skull.fcsv", "spec2_skull.fcsv", "spec3_skull.fcsv", etc.
# dir <- "~/Documents/LMs")
# skull_array <- fcsv2array(dir = dir, pattern="_skull", string_del="_skull")
  
}
\author{
Marta Vidal-Garcia
}
