% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers_percentile.R
\name{plotOutliers_percentile}
\alias{plotOutliers_percentile}
\title{plotOutliers_percentile}
\usage{
plotOutliers_percentile(A, groups = NULL, inspect.outliers = FALSE, percentile, save.plot = FALSE)
}
\description{
This function has been adapted from the plotOutliers function in geomorph, and it has been modified to select for percentiles and save plots
}
\details{
This function is highly based on the plotOutliers function in geomorph (Adams et al., 2018), and it has been modified to select for percentiles and save plots (including a histogram). It will check for outliers at the selected percentile (e.g. 95).
}
\examples{
   species <- c("sp_A", "sp_A", "sp_B", "sp_A", "sp_A", "sp_B", "sp_B", "sp_A", "sp_A", "sp_A", "sp_A")
    plotOutliers_percentile(GPA$coords, percentile = 0.95) 
    plotOutliers_percentile(GPA$coords, percentile = 0.99, groups = species, save.plot = TRUE)
  
}
\references{
Adams, D. C., M. L. Collyer, and A. Kaliontzopoulou. 2018. Geomorph: Software for geometric morphometric analysis. R package version 3.0.6. Available at http://CRAN.R-project.org/package=geomorph.
}
\author{
Marta Vidal-Garcia
}
