% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp2array.R
\name{pp2array}
\alias{pp2array}
\title{pp2array}
\usage{
pp2array(dir = NULL, ID = NULL, string_del = NULL, save.txt = FALSE)
}
\arguments{
\item{dir}{Optional argument. Directory where the *.pp files are. Default (NULL) is the current working directory.}

\item{ID}{Optional argument for specimens IDs. The default (NULL) is the file names.}

\item{string_del}{Optional argument. Indicates pattern in the specimens names to be deleted. Use only if ID = NULL.}

\item{save.txt}{Optional argument. Indicates whether to also save *.txt files with the landmark coordinates.}
}
\value{
This function returns an array of dimensions (p, k, n), in which p is the number of landmarks, k = 3 (number of dimensions), and n is the number of specimens.
}
\description{
This function inputs *pp files into an array of dimensions p, k, n
}
\examples{
# array <- pp2array()
# If we have other *.pp files in another directory and with the suffix "skull"
# For example: "spec1_skull.pp", "spec2_skull.pp", "spec3_skull.pp", etc.
# dir <- "~/Documents/skull_LMs")
# skull_array <- pp2array(dir = skull_dir, string_del="skull")

}
\author{
Marta Vidal-Garcia
}
