% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp2lm.R
\name{pp2lm}
\alias{pp2lm}
\title{pp2lm}
\usage{
pp2lm(file = file)
}
\arguments{
\item{file}{#'}
}
\value{
This function returns  matrix of n rows (number of landmarks) and x, y, z columns.
}
\description{
This function inputs a single *pp file into a matrix of n rows and x, y, z columns
}
\examples{
# LMs <- pp2lm("./test_LMs.pp")
}
\author{
Marta Vidal-Garcia
}
