% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2array.R
\name{tag2array}
\alias{tag2array}
\title{tag2array}
\usage{
tag2array(dir = NULL, ID = NULL, string_del = NULL)
}
\arguments{
\item{dir}{Optional argument. Directory where the *.ascii files are. Default (NULL) is the current working directory.}

\item{ID}{Optional argument for specimens IDs. The default (NULL) is the file names.}

\item{string_del}{Optional argument. Indicates pattern in the specimens names to be deleted. Use only if ID = NULL.}
}
\value{
This function returns an array of dimensions (p, k, n), in which p is the number of landmarks, k = 3 (number of dimensions), and n is the number of specimens.
}
\description{
This function inputs *tag files into an array of dimensions p, k, n
}
\examples{
# array <- tag2array()
# If we have other *.tag files in another directory and with the suffix "skull"
# For example: "spec1_skull.tag", "spec2_skull.tag", "spec3_skull.tag", etc.
# dir <- "~/Documents/skull_LMs")
# skull_array <- tag2array(dir = skull_dir, string_del="skull")
  
}
\author{
Marta Vidal-Garcia
}
