% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2lm.R
\name{tag2lm}
\alias{tag2lm}
\title{tag2lm}
\usage{
tag2lm(dir = NULL, ID = NULL, string_del = NULL)
}
\arguments{
\item{\item{dir}{}{}
}
\description{
This function inputs *tag files into an array of dimensions p, k, n
}
\details{
This function inputs *tag files into an array of dimensions (p, k, n), in which p is the number of landmarks, k = 3 (number of dimensions), and n is the number of specimens.
}
\examples{
# array <- tag2lm()
# If we have *.tag file in another directory and with the suffix "skull"
# For example: "spec1_skull.tag", "spec2_skull.tag", "spec3_skull.tag", etc.
dir <- "~/Documents/skull_LMs")
# skull_array <- tag2lm(dir = skull_dir, string_del="skull")
  
}
\author{
Marta Vidal-Garcia
}
