% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.tag.R
\name{write.tag}
\alias{write.tag}
\title{write.tag}
\usage{
write.tag(mat, file)
}
\arguments{
\item{mat}{xy matrix with three columns (x, y, z) and the landmarks ordered by row}

\item{file}{name of the tag file to save}
}
\value{
This function saves a *.tag file from a matrix
}
\description{
This function inputs *tag files into an array of dimensions p, k, n
}
\examples{
# mat <- matrix(c(1,2,3, 4,5,6, 7,8,9, 10,11,12), nrow = 4, ncol = 3)
# write.tag(mat, "my_matrix.tag")
  
}
\author{
Marta Vidal-Garcia
}
