% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.volume.R
\name{morpho.volume}
\alias{morpho.volume}
\title{morpho.volume}
\usage{
morpho.volume(x, groups)
}
\arguments{
\item{x}{Dataset with three columns (3 dimensions in the morphospace)}

\item{groups}{An optional factor defining groups}
}
\value{
This function returns a list with two objects. The first object is the volume for the whole dataset. If groups have been selected, then it also returns a data.frame with the volumes by group and the percentage of the total volume occupied in the morphospace.
}
\description{
This function returns the volume (and volumes by groups) of a morphospace.
}
\examples{
   # x <- pc.scores[,1:3] # first three PC scores
   # groups <- c("sp_A", "sp_A", "sp_B", "sp_A", "sp_A", "sp_B", "sp_B", "sp_A", "sp_A", "sp_A", "sp_A")
   # names(groups) <- row.names(x)
   # morpho.volume(x, groups) 
  
}
\author{
Marta Vidal-Garcia
}
