% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho.volume.R
\name{morpho.volume}
\alias{morpho.volume}
\title{morpho.volume}
\usage{
morpho.volume(x, groups)
}
\description{
This function returns the volume (and volumes by groups) of a morphospace.
}
\details{
This function returns the volume (and volumes by groups) occupied in the morphospace.
}
\examples{
    x <- pc.scores[,1:3] # first three PC scores
    groups <- c("sp_A", "sp_A", "sp_B", "sp_A", "sp_A", "sp_B", "sp_B", "sp_A", "sp_A", "sp_A", "sp_A")
    names(groups) <- row.names(x)
    morpho.volume(x, groups) 
  
}
\author{
Marta Vidal-Garcia
}
