% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2lm.R
\name{tag2lm}
\alias{tag2lm}
\title{tag2lm}
\usage{
tag2lm(file)
}
\arguments{
\item{file}{MINC *.tag file to import.}
}
\value{
This function returns matrix with the landmark coordinates (x, y, z)
}
\description{
This function inputs a single *tag file into an matrix with the landmark coordinates (x, y, z)
}
\examples{
# specimen1 <- tag2lm("specimen1.tag")
  
}
\author{
Marta Vidal-Garcia
}
