% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.tag.R
\name{write.tag}
\alias{write.tag}
\title{write.tag}
\usage{
write.tag(mat, file)
}
\arguments{
\item{mat}{xy matrix with three columns (x, y, z) and the landmarks ordered by row}

\item{file}{name of the tag file to save}
}
\value{
This function saves a *.tag file from a matrix
}
\description{
This function exports a matrix to a *.tag file
}
\examples{
# mat <- matrix(1:12), nrow = 4, ncol = 3)
# write.tag(mat, "my_matrix.tag")
  
}
\author{
Marta Vidal-Garcia
}
